content/canvas/test/test_2d.path.isPointInPath.subpath.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.isPointInPath.subpath</title>
<!-- Testing: isPointInPath() uses the current path, not just the subpath -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.rect(0, 0, 20, 20);
ctx.beginPath();
ctx.rect(20, 0, 20, 20);
ctx.closePath();
ctx.rect(40, 0, 20, 20);
ok(ctx.isPointInPath(10, 10) === false, "ctx.isPointInPath(10, 10) === false");
ok(ctx.isPointInPath(30, 10) === true, "ctx.isPointInPath(30, 10) === true");
ok(ctx.isPointInPath(50, 10) === true, "ctx.isPointInPath(50, 10) === true");

SimpleTest.finish();

});
</script>
